refactor: left,right div 페이지 구분
